Changed code for UML diagram correctness and more readability

+++ src/classes/writer_class.py
@@ -12,27 +12,28 @@ class Writer:
     stick_count = 0
     obstacle_count = 0
 
-    def __init__(self):
+    def __init__(self, board):
         """
         Initialization function that deletes csv file if it exists.
         """
+        self.board = board
         self.remove_csv_file()
 
     def remove_csv_file(self):
         """
         Removes csv file in folder
         """
-        os.remove('results.csv')
+        if os.path.isfile('results.csv'):
+            os.remove('results.csv')
 
-    def write_to_csv(self, board):
+    def write_to_csv(self):
         """
-        It writes information about exact round to the csv file by appends mode
-        :param Board board: Is used to get information from board
+        It writes information about exact round to the csv file by append mode
         """
         round_count = Round.count
-        food_count = len(board.get_thing('food'))
-        stick_count = len(board.get_thing('stick'))
-        obstacle_count = len(board.get_thing('obstacle'))
+        food_count = len(self.board.get_thing('food'))
+        stick_count = len(self.board.get_thing('stick'))
+        obstacle_count = len(self.board.get_thing('obstacle'))
 
         with open('results.csv', mode='a') as result_file:
             result_writer = csv.writer(result_file, delimiter=',', quotechar='"', quoting=csv.QUOTE_MINIMAL)
